Created FishLibrary.kt to store fish paths
